+++ test/c/test_target.c
@@ -432,7 +432,7 @@ int test_target_add_instruction(void) {
         uint32_t q[1] = {i};
         qk_target_entry_add_property(meas, q, 1, 1e-6, 1e-4);
     }
-    uint32_t num_meas = qk_target_entry_num_properties(meas);
+    size_t num_meas = qk_target_entry_num_properties(meas);
     if (num_meas != 3) {
         printf("Expected 3 measurement entries but got: %u", num_meas);
         result = EqualityError;
@@ -468,7 +468,7 @@ int test_target_add_instruction(void) {
         uint32_t q[1] = {i};
         qk_target_entry_add_property(reset, q, 1, 2e-6, 2e-4);
     }
-    uint32_t num_reset = qk_target_entry_num_properties(reset);
+    size_t num_reset = qk_target_entry_num_properties(reset);
     if (num_reset != 3) {
         printf("Expected 3 reset entries but got: %u", num_reset);
         result = EqualityError;
